requirements.txt
yt-dlp
pytest
pytest-mock
faster-whisper
chromadb
langchain-text-splitters
litellm
eel
fastapi
uvicorn
pywebview
playwright
python-multipart
